Move IExtractor & IParser to recognizers-base
