+++ JavaScript/packages/recognizers-base/src/recognizers-text-base.ts
@@ -1,3 +1,5 @@
 export { Culture, CultureInfo } from "./culture";
 export { IModel, ModelResult, ModelContainer } from "./models";
 export { IRecognizer, Recognizer } from "./recognizer";
+export { IExtractor, ExtractResult } from "./extractors";
+export { IParser, ParseResult } from "./parsers";
