+++ JavaScript/packages/recognizers-date-time/src/dateTime/baseDateTimePeriod.ts
@@ -1,5 +1,6 @@
+import { IExtractor, ExtractResult } from "recognizers-text-base";
 import { Constants, TimeTypeConstants } from "./constants";
-import { IExtractor, ExtractResult, BaseNumberExtractor, RegExpUtility, StringUtility } from "recognizers-text-number"
+import { BaseNumberExtractor, RegExpUtility, StringUtility } from "recognizers-text-number"
 import { BaseDateExtractor, BaseDateParser } from "./baseDate"
 import { BaseTimeExtractor, BaseTimeParser } from "./baseTime"
 import { BaseDateTimeExtractor, BaseDateTimeParser, IDateTimeExtractor } from "./baseDateTime"
