+++ JavaScript/packages/recognizers-date-time/src/dateTime/chinese/baseDateTime.ts
@@ -1,4 +1,5 @@
-import { IExtractor, ExtractResult, StringUtility, Match, RegExpUtility } from "recognizers-text-number";
+import { IExtractor, ExtractResult } from "recognizers-text-base";
+import { StringUtility, Match, RegExpUtility } from "recognizers-text-number";
 import { IDateTimeExtractor } from "../baseDateTime"
 import { Constants, TimeTypeConstants } from "../constants"
 import { Token } from "../utilities";
