+++ JavaScript/packages/recognizers-date-time/src/dateTime/chinese/dateTimeConfiguration.ts
@@ -1,4 +1,5 @@
-import { IExtractor, ExtractResult, BaseNumberParser, BaseNumberExtractor, RegExpUtility, StringUtility } from "recognizers-text-number"
+import { IExtractor, ExtractResult } from "recognizers-text-base";
+import { BaseNumberParser, BaseNumberExtractor, RegExpUtility, StringUtility } from "recognizers-text-number"
 import { Constants, TimeTypeConstants } from "../constants";
 import { IDateTimeExtractor, IDateTimeExtractorConfiguration, BaseDateTimeExtractor, IDateTimeParserConfiguration, BaseDateTimeParser } from "../baseDateTime";
 import { BaseDurationExtractor, BaseDurationParser } from "../baseDuration"
