+++ JavaScript/packages/recognizers-date-time/src/dateTime/chinese/dateTimePeriodConfiguration.ts
@@ -1,4 +1,5 @@
-import { ChineseNumberParserConfiguration, AgnosticNumberParserFactory, AgnosticNumberParserType, IExtractor, ExtractResult, IParser, BaseNumberParser, BaseNumberExtractor, RegExpUtility, StringUtility, ChineseCardinalExtractor } from "recognizers-text-number"
+import { IExtractor, IParser, ExtractResult } from "recognizers-text-base";
+import { ChineseNumberParserConfiguration, AgnosticNumberParserFactory, AgnosticNumberParserType, BaseNumberParser, BaseNumberExtractor, RegExpUtility, StringUtility, ChineseCardinalExtractor } from "recognizers-text-number"
 import { Constants, TimeTypeConstants } from "../constants";
 import { IDateTimePeriodExtractorConfiguration, BaseDateTimePeriodExtractor, IDateTimePeriodParserConfiguration, BaseDateTimePeriodParser } from "../baseDateTimePeriod";
 import { BaseDurationExtractor, BaseDurationParser } from "../baseDuration"
