+++ JavaScript/packages/recognizers-date-time/src/dateTime/chinese/durationConfiguration.ts
@@ -1,4 +1,5 @@
-import { ExtractResult, RegExpUtility, CultureInfo, Culture, BaseNumberExtractor, BaseNumberParser } from "recognizers-text-number";
+import { ExtractResult } from "recognizers-text-base";
+import { RegExpUtility, CultureInfo, Culture, BaseNumberExtractor, BaseNumberParser } from "recognizers-text-number";
 import { NumberWithUnitExtractor, ChineseNumberWithUnitExtractorConfiguration, NumberWithUnitParser, ChineseNumberWithUnitParserConfiguration, UnitValue } from "recognizers-text-number-with-unit";
 import { BaseDateTimeExtractor } from "./baseDateTime";
 import { IDurationParserConfiguration, BaseDurationParser } from "../baseDuration";
