+++ JavaScript/packages/recognizers-date-time/src/dateTime/chinese/holidayConfiguration.ts
@@ -1,5 +1,6 @@
+import { IExtractor, IParser, ExtractResult } from "recognizers-text-base";
 import { IHolidayExtractorConfiguration, BaseHolidayParserConfiguration, BaseHolidayParser } from "../baseHoliday"
-import { RegExpUtility, ExtractResult, Match, IExtractor, IParser, StringUtility, ChineseIntegerExtractor, AgnosticNumberParserFactory, AgnosticNumberParserType, ChineseNumberParserConfiguration } from "recognizers-text-number";
+import { RegExpUtility, Match, StringUtility, ChineseIntegerExtractor, AgnosticNumberParserFactory, AgnosticNumberParserType, ChineseNumberParserConfiguration } from "recognizers-text-number";
 import { Constants as NumberConstants } from "recognizers-text-number"
 import { DateUtils, FormatUtil, DateTimeResolutionResult, StringMap } from "../utilities";
 import { ChineseDateTime } from "../../resources/chineseDateTime";
