+++ JavaScript/packages/recognizers-date-time/src/dateTime/chinese/timePeriodConfiguration.ts
@@ -1,4 +1,5 @@
-import { ExtractResult, RegExpUtility, CultureInfo, Culture } from "recognizers-text-number";
+import { ExtractResult } from "recognizers-text-base";
+import { RegExpUtility, CultureInfo, Culture } from "recognizers-text-number";
 import { NumberWithUnitExtractor, ChineseNumberWithUnitExtractorConfiguration } from "recognizers-text-number-with-unit";
 import { BaseDateTimeExtractor, DateTimeExtra, TimeResult, TimeResolutionUtils } from "./baseDateTime";
 import { Constants, TimeTypeConstants } from "../constants"
