+++ JavaScript/packages/recognizers-date-time/src/dateTime/french/timeConfiguration.ts
@@ -1,4 +1,5 @@
-import { RegExpUtility, IExtractor } from "recognizers-text-number";
+import { IExtractor } from "recognizers-text-base";
+import { RegExpUtility } from "recognizers-text-number";
 import { ITimeExtractorConfiguration, ITimeParserConfiguration } from "../baseTime";
 import { FrenchDateTime } from "../../resources/frenchDateTime";
 import { BaseDurationExtractor } from "../baseDuration";
