+++ JavaScript/packages/recognizers-date-time/src/dateTime/models.ts
@@ -1,5 +1,5 @@
-import { IModel, ModelResult } from "recognizers-text-base";
-import { IExtractor, ParseResult, FormatUtility } from "recognizers-text-number";
+import { IModel, ModelResult, IExtractor, ParseResult } from "recognizers-text-base";
+import { FormatUtility } from "recognizers-text-number";
 import { IDateTimeParser, DateTimeParseResult } from "./parsers";
 
 export class DateTimeModelResult extends ModelResult {
