+++ JavaScript/packages/recognizers-date-time/src/dateTime/spanish/datePeriodConfiguration.ts
@@ -1,4 +1,5 @@
-import { BaseNumberExtractor, IExtractor, SpanishIntegerExtractor, RegExpUtility, IParser } from "recognizers-text-number";
+import { IExtractor, IParser } from "recognizers-text-base";
+import { BaseNumberExtractor, SpanishIntegerExtractor, RegExpUtility } from "recognizers-text-number";
 import { IDatePeriodExtractorConfiguration, IDatePeriodParserConfiguration } from "../baseDatePeriod";
 import { BaseDateExtractor, BaseDateParser } from "../baseDate";
 import { BaseDurationExtractor, BaseDurationParser } from "../baseDuration";
