+++ JavaScript/packages/recognizers-date-time/src/dateTime/utilities.ts
@@ -1,4 +1,5 @@
-import { IExtractor, ExtractResult, RegExpUtility } from "recognizers-text-number"
+import { IExtractor, ExtractResult } from "recognizers-text-base";
+import { RegExpUtility } from "recognizers-text-number"
 import { IDateTimeParser, DateTimeParseResult } from "../dateTime/parsers"
 import { TimeTypeConstants } from "../dateTime/constants"
 import { IDateTimeExtractor } from "./baseDateTime";
