+++ JavaScript/packages/recognizers-date-time/src/recognizers-text-date-time.ts
@@ -49,7 +49,6 @@ export { FrenchSetExtractorConfiguration, FrenchSetParserConfiguration } from ".
 export { FrenchDurationExtractorConfiguration, FrenchDurationParserConfiguration } from "./dateTime/french/durationConfiguration";
 export { FrenchHolidayExtractorConfiguration, FrenchHolidayParserConfiguration } from "./dateTime/french/holidayConfiguration";
 export { FrenchMergedExtractorConfiguration, FrenchMergedParserConfiguration } from "./dateTime/french/mergedConfiguration";
-
 export { ChineseDurationExtractor, ChineseDurationParser } from "./dateTime/chinese/durationConfiguration";
 export { ChineseTimeExtractor, ChineseTimeParser } from "./dateTime/chinese/timeConfiguration";
 export { ChineseTimePeriodExtractor, ChineseTimePeriodParser } from "./dateTime/chinese/timePeriodConfiguration";
