+++ JavaScript/packages/recognizers-number-with-unit/src/numberWithUnit/parsers.ts
@@ -1,4 +1,5 @@
-import { CultureInfo, IParser, ParseResult, IExtractor, ExtractResult } from "recognizers-text-number";
+import { IExtractor, ExtractResult, IParser, ParseResult } from "recognizers-text-base";
+import { CultureInfo } from "recognizers-text-number";
 import last = require("lodash.last");
 
 export class UnitValue {
