+++ JavaScript/packages/recognizers-number/src/number/chinese/parserConfiguration.ts
@@ -1,4 +1,5 @@
-import { INumberParserConfiguration, ParseResult } from "../parsers";
+import { ParseResult } from "recognizers-text-base";
+import { INumberParserConfiguration } from "../parsers";
 import { CultureInfo, Culture } from "../../culture";
 import { ChineseNumeric } from "../../resources/chineseNumeric";
 import { RegExpUtility } from "../../utilities"
