+++ JavaScript/packages/recognizers-number/src/number/french/parserConfiguration.ts
@@ -1,4 +1,5 @@
-import { INumberParserConfiguration, ParseResult } from "../parsers";
+import { ParseResult } from "recognizers-text-base";
+import { INumberParserConfiguration } from "../parsers";
 import { CultureInfo, Culture } from "../../culture";
 import { FrenchNumeric } from "../../resources/frenchNumeric";
 import { RegExpUtility } from "../../utilities"
