+++ JavaScript/packages/recognizers-number/src/recognizers-text-number.ts
@@ -7,10 +7,10 @@ export { SpanishNumeric } from "./resources/spanishNumeric";
 export { FrenchNumeric } from "./resources/frenchNumeric";
 export { ChineseNumeric } from "./resources/chineseNumeric";
 export { Constants } from "./number/constants";
-export { IExtractor, ExtractResult, RegExpValue, BaseNumberExtractor, BasePercentageExtractor } from "./number/extractors";
+export { RegExpValue, BaseNumberExtractor, BasePercentageExtractor } from "./number/extractors";
 export { NumberMode, LongFormatType, AbstractNumberModel, NumberModel, OrdinalModel, PercentModel } from "./number/models";
 export { AgnosticNumberParserType, AgnosticNumberParserFactory } from "./number/agnosticNumberParser";
-export { IParser, ParseResult, INumberParserConfiguration, BaseNumberParser, BasePercentageParser } from "./number/parsers";
+export { INumberParserConfiguration, BaseNumberParser, BasePercentageParser } from "./number/parsers";
 export { EnglishCardinalExtractor, EnglishDoubleExtractor, EnglishFractionExtractor, EnglishIntegerExtractor, EnglishNumberExtractor, EnglishOrdinalExtractor, EnglishPercentageExtractor } from "./number/english/extractors";
 export { EnglishNumberParserConfiguration } from "./number/english/parserConfiguration";
 export { SpanishCardinalExtractor, SpanishDoubleExtractor, SpanishFractionExtractor, SpanishIntegerExtractor, SpanishNumberExtractor, SpanishOrdinalExtractor, SpanishPercentageExtractor } from "./number/spanish/extractors";
